split int_vol dtype tests, update codecov
- split the testing for integral volume for int and floats respectively
- add coverage for "features"
- add "..." to ignore (whic should resolve the missing coverage for
overload methods

+++ afids_utils/features/regRF.py
@@ -11,19 +11,20 @@ from numpy.typing import NDArray
 
 @overload
 def integral_volume(nii_img: NDArray[np.int_]) -> NDArray[np.int_]:
-    ...
+    ...  # pragma: no cover
 
 
 @overload
 def integral_volume(nii_img: NDArray[np.float_]) -> NDArray[np.float_]:
-    ...
+    ...  # pragma: no cover
 
 
 def integral_volume(nii_img: NDArray[np.int_ | np.float_]):
     """Compute zero-padded (resampled) volume of image"""
     iv_img = nii_img.cumsum(0).cumsum(1).cumsum(2)
     iv_zeropad = np.zeros(
-        (iv_img.shape[0] + 1, iv_img.shape[1] + 1, iv_img.shape[2] + 1)
+        (iv_img.shape[0] + 1, iv_img.shape[1] + 1, iv_img.shape[2] + 1),
+        dtype=iv_img.dtype,
     )
     iv_zeropad[1:, 1:, 1:] = iv_img
